Second fix in mergin problems

+++ Server/src/main/java/org/gluu/oxauth/service/AuthorizeService.java
@@ -28,14 +28,7 @@ import org.gluu.oxauth.ciba.CIBAPushErrorService;
 import org.gluu.oxauth.model.authorize.AuthorizeErrorResponseType;
 import org.gluu.oxauth.model.authorize.AuthorizeRequestParam;
 import org.gluu.oxauth.model.ciba.PushErrorResponseType;
-import org.gluu.oxauth.model.common.AuthorizationGrantList;
-import org.gluu.oxauth.model.common.CibaRequestCacheControl;
-import org.gluu.oxauth.model.common.CibaRequestStatus;
-import org.gluu.oxauth.model.common.Prompt;
-import org.gluu.oxauth.model.common.ResponseMode;
-import org.gluu.oxauth.model.common.ResponseType;
-import org.gluu.oxauth.model.common.SessionId;
-import org.gluu.oxauth.model.common.User;
+import org.gluu.oxauth.model.common.*;
 import org.gluu.oxauth.model.configuration.AppConfiguration;
 import org.gluu.oxauth.model.error.ErrorResponseFactory;
 import org.gluu.oxauth.model.registration.Client;
@@ -309,12 +302,10 @@ public class AuthorizeService {
         String userCode = sessionAttribute.get(AuthorizeRequestParam.USER_CODE);
         DeviceAuthorizationCacheControl cacheData = deviceAuthorizationService.getDeviceAuthzByUserCode(userCode);
 
-        if (cacheData != null) {
-            if (cacheData.getStatus() == DeviceAuthorizationStatus.PENDING) {
-                cacheData.setStatus(DeviceAuthorizationStatus.DENIED);
-                deviceAuthorizationService.saveInCache(cacheData, true, false);
-                deviceAuthorizationService.removeDeviceAuthRequestInCache(userCode, null);
-            }
+        if (cacheData != null && cacheData.getStatus() == DeviceAuthorizationStatus.PENDING) {
+            cacheData.setStatus(DeviceAuthorizationStatus.DENIED);
+            deviceAuthorizationService.saveInCache(cacheData, true, false);
+            deviceAuthorizationService.removeDeviceAuthRequestInCache(userCode, null);
         }
     }
 }
